Move usage controllers and webhooks under 'protection' group
